+++ src/components/Jobs/RunningJobs/tempData.js
@@ -1,0 +1,28 @@
+const tempData = [{
+  id: 'de96f7a1-8706-42e9-ba86-c44c44592511',
+  hrId: 3,
+  parentJobId: 'de96f7a1-8706-42e9-ba86-c44c44592511',
+  subordinationType: 'PARENT_SINGLE',
+  jobProfileInfo: {
+    id: '558dede0-515d-11e9-8647-d663bd873d93',
+    name: 'default',
+    dataType: 'MARC',
+  },
+  sourcePath: 'EBOOKS1.mrc',
+  fileName: 'EBOOKS1.mrc',
+  runBy: {
+    firstName: 'John',
+    lastName: 'Doe',
+  },
+  progress: {
+    jobExecutionId: 'de96f7a1-8706-42e9-ba86-c44c44592511',
+    current: 21650,
+    total: 45265,
+  },
+  startedDate: '2020-01-27T10:10:18.743+0000',
+  status: 'PARSING_IN_PROGRESS',
+  uiStatus: 'RUNNING',
+  userId: '80328998-31a1-519f-a1b3-ba81f7ce6f8c',
+}];
+
+export default tempData;