+++ test/bigtest/interactors/index.js
@@ -1,3 +1,5 @@
 export * from './appDisplayInteractor';
 export * from './settingsDisplayInteractor';
 export * from './queryFileUploaderInteractor';
+export * from './jobs/JobsInteractor';
+export * from './jobs/RunningJobInteractor';